WIP example app improvements
