+++ example/lib/chart_with_labels.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:spider_chart/spider_chart.dart';
+
+class ChartWithLabels extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Chart With Labels'),
+      ),
+      body: Center(
+        child: Container(
+          width: 300,
+          height: 300,
+          child: SpiderChart(
+              data: [
+                7,
+                5,
+                10,
+                7,
+                4,
+              ],
+              maxValue: 10,
+              colors: <Color>[
+                Colors.red,
+                Colors.green,
+                Colors.blue,
+                Colors.yellow,
+                Colors.indigo,
+              ],
+              labels: <String>[
+                "label 1",
+                "label 2",
+                "label 3",
+                "label 4",
+                "label 5",
+              ]),
+        ),
+      ),
+    );
+  }
+}
